Add back the logging
